[#47] Move all entities to one directory
