+++ package.json
@@ -86,11 +86,14 @@
     ],
     "rootDir": "./",
     "moduleNameMapper": {
-      "@shikicinema": [
+      "@lib-shikicinema": [
         "<rootDir>/lib/shikicinema/index.ts"
       ],
-      "@utils/(.*)": [
+      "@app-utils/(.*)": [
         "<rootDir>/src/utils/$1"
+      ],
+      "@app-entities": [
+        "<rootDir>/src/entities/index.ts"
       ],
       "dist/(.*)": [
         "<rootDir>/dist/$1"
