--- src/config/entities.config.ts
@@ -1,7 +1,0 @@
-import { UserEntity } from '../services/user/user.entity';
-import { UploaderEntity } from '../services/uploader/uploader.entity';
-import { UploadTokenEntity } from '../services/upload-tokens/upload-token.entity';
-import { SessionEntity } from '../services/session/session.entity';
-
-// TODO: change this hardcode
-export default [ UserEntity, UploaderEntity, UploadTokenEntity, SessionEntity ];
