+++ src/entities/index.ts
@@ -1,0 +1,12 @@
+import { UserEntity } from './user';
+import { UploaderEntity } from './uploader';
+import { UploadTokenEntity } from './upload-token';
+import { SessionEntity } from './session';
+
+export {SessionEntity} from './session';
+export {UploadTokenEntity} from './upload-token';
+export {UploaderEntity} from './uploader';
+export {UserEntity} from './user';
+
+
+export const entities =  [UserEntity, UploaderEntity, UploadTokenEntity, SessionEntity];
