+++ src/routes/api/admin/user/dto/AdminUserInfo.dto.ts
@@ -1,6 +1,6 @@
-import { AdminUser, Role } from '@shikicinema';
+import { AdminUser, Role } from '@lib-shikicinema';
 import { Exclude, Expose } from 'class-transformer';
-import { TransformRoles } from '@utils/class-transform.utils';
+import { TransformRoles } from '@app-utils/class-transform.utils';
 
 @Exclude()
 export class AdminUserInfo implements AdminUser {
