+++ src/routes/auth/auth.service.spec.ts
@@ -8,14 +8,11 @@ import {
 import { Test, TestingModule } from '@nestjs/testing';
 import * as bcrypt from 'bcrypt';
 import { UserService } from '../../services/user/user.service';
-import { UserEntity } from '../../services/user/user.entity';
+import { SessionEntity, UploaderEntity, UploadTokenEntity, UserEntity } from '@app-entities';
 import { SessionService } from '../../services/session/session.service';
-import { SessionEntity } from '../../services/session/session.entity';
 import { ShikimoriClient } from '../../services/shikimori/shikimori.client';
-import { UploadTokenEntity } from '../../services/upload-tokens/upload-token.entity';
 import { UploadTokensService } from '../../services/upload-tokens/upload-tokens.service';
 import { ShikimoriUser } from '../../services/shikimori/dto/ShikimoriUser.dto';
-import { UploaderEntity } from '../../services/uploader/uploader.entity';
 import { PgSharedService } from '../../services/postgres/postgres.service';
 import { UploaderService } from '../../services/uploader/uploader.service';
 import { AuthService } from './auth.service';
