+++ src/routes/auth/upload-token.strategy.ts
@@ -2,7 +2,7 @@ import { PassportStrategy } from '@nestjs/passport';
 import { ForbiddenException, Injectable, UnauthorizedException } from '@nestjs/common';
 import { Strategy as BearerStrategy } from 'passport-http-bearer';
 import { UploadTokensService } from '../../services/upload-tokens/upload-tokens.service';
-import { UploadTokenEntity } from '../../services/upload-tokens/upload-token.entity';
+import { UploadTokenEntity } from '@app-entities';
 
 @Injectable()
 export class UploadTokenStrategy extends PassportStrategy(BearerStrategy) {
