+++ src/services/session/session.shared.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SessionService } from './session.service';
-import { SessionEntity } from './session.entity';
+import { SessionEntity } from '@app-entities';
 
 @Module({
   providers: [SessionService],
