+++ src/services/user/dto/GetUsers.dto.ts
@@ -1,7 +1,7 @@
-import { GetUsersRequest, Role } from '@shikicinema';
+import { GetUsersRequest, Role } from '@lib-shikicinema';
 import { IsArray, IsDate, IsEmail, IsEnum, IsInt, IsOptional, IsString, Length, Min } from 'class-validator';
 import { Exclude, Expose, Type } from 'class-transformer';
-import { TransformDate, TransformNullableString, TransformRoles } from '@utils/class-transform.utils';
+import { TransformDate, TransformNullableString, TransformRoles } from '@app-utils/class-transform.utils';
 
 @Exclude()
 export class GetUsers implements GetUsersRequest {
