+++ src/utils/class-transform.utils.ts
@@ -1,6 +1,6 @@
 import { Transform } from 'class-transformer';
 import { TransformOptions } from 'class-transformer/metadata/ExposeExcludeOptions';
-import { Role } from '@shikicinema';
+import { Role } from '@lib-shikicinema';
 
 export function TransformNullableString(options?: TransformOptions) {
   return Transform((value) => {
