+++ tsconfig.json
@@ -13,8 +13,9 @@
         "baseUrl": "./",
         "incremental": true,
         "paths": {
-            "@shikicinema": [ "<rootDir>/../lib/shikicinema"],
-            "@utils/*": ["<rootDir>/../src/utils/*"],
+            "@lib-shikicinema": [ "<rootDir>/../lib/shikicinema"],
+            "@app-utils/*": ["<rootDir>/../src/utils/*"],
+            "@app-entities": ["<rootDir>/../src/entities"],
             "dist/*": ["<rootDir>/../dist/*"]
         }
     }
